REFACTOR: Address code review findings for CI/CD setup
- Extract shared LLM prompts, model constant, and JSON parser into workers/shared.js
- Extract duplicated route handlers into routes/extractions.js with factory pattern
- Simplify api/index.js to use shared routes (129 → 35 lines)
- Lazy-require OpenAI SDK to reduce serverless cold-start time
- Configure pg Pool for serverless (max 1 conn, timeouts)
- Consolidate CI workflow into single job (was 2 redundant jobs)
- Consolidate deploy workflow with environment detection (was 2 near-identical jobs)
- Remove broken db:migrate:prod script (handled by migrate.yml workflow)
- Replace body-parser with built-in express.json()
- Fix inconsistent JSON fence stripping regex
