+++ db/index.js
@@ -10,8 +10,15 @@ const { drizzle } = require('drizzle-orm/node-postgres');
 const { Pool } = require('pg');
 const schema = require('./schema');
 
+const isServerless = !!process.env.VERCEL;
+
 const pool = new Pool({
   connectionString: process.env.DATABASE_URL || 'postgresql://claude:claude@localhost:5432/stonie_notes',
+  ...(isServerless && {
+    max: 1,
+    idleTimeoutMillis: 20000,
+    connectionTimeoutMillis: 5000,
+  }),
 });
 
 const db = drizzle(pool, { schema });
